+++ src/components/window/index.ts
@@ -1,28 +1,33 @@
-export { WindowControl } from './window-control'
+export { Window } from './window'
+export type { WindowOwnProps, WindowProps, WindowRootProps } from './window.types'
+
 export {
 	resolveWindowControlAnimation,
 	WINDOW_CONTROL_DEFAULT_GRABBER_HOVER_OPACITY,
 	WINDOW_CONTROL_DEFAULT_GRABBER_REST_OPACITY,
+	WINDOW_CONTROL_DEFAULT_GRABBER_SHRINK_SCALE,
 	WINDOW_CONTROL_DEFAULT_GRABBER_TRANSITION,
 	WINDOW_CONTROL_DEFAULT_HIDDEN_OPACITY,
 	WINDOW_CONTROL_DEFAULT_HIDDEN_SCALE,
 	WINDOW_CONTROL_DEFAULT_REVEAL_TRANSITION,
-} from './window-control.animation'
-export { WINDOW_CONTROL_SIZE } from './window-control.constants'
+} from './window.control.animation'
+export type { WindowControlSide } from './window.control.context'
 export {
 	windowControlButtonVariants,
 	windowControlGrabberVariants,
+	windowControlSlotVariants,
 	windowControlVariants,
-} from './window-control.styles'
+} from './window.control.styles'
 export type {
 	WindowControlAnimation,
+	WindowControlButtonOwnProps,
 	WindowControlButtonProps,
-	WindowControlCloseProps,
 	WindowControlGrabberAnimation,
 	WindowControlGrabberProps,
 	WindowControlProps,
 	WindowControlRevealAnimation,
 	WindowControlRootOwnProps,
 	WindowControlRootProps,
-	WindowControlShareProps,
-} from './window-control.types'
+	WindowControlSlotOwnProps,
+	WindowControlSlotProps,
+} from './window.control.types'